First comment
This is my first project using Python. I started learning it last month, so to get started, you may need to have Visual Studio or whatever you use to program and have the top import things for docx downloaded. This is a simple gym workout tracker. The first question it will ask is the exercise name. After you set it up. Then you put the lbs you want to go up by then (leaving it blank; it goes up by 5). Then, write out the weight you were using for the exercises for each set (recommended 3 sets) (also essential to leave space for each set you would like to add). Then, do the same thing for the following question: the reps you got for the given reps you did for each set. Then, it would calculate your exercise weight based on my hypertrophy training. For the first set, you must get 12 reps or more to increase weight. For the second, you will need to get 10 reps or greater to increase the weight, and so on. This is the same in decreasing weight, so with a first set, if you do 6 reps or less, it will reduce the weight. The same thing applies, decreasing by 2 reps with each following set. After that, it will repeat the loop, asking for the next exercise until you put in all the exercises, and when you are done, put in "end" in the first question, asking for the name of the exercise. What, then, it will create a Word file with your exercises ready to print. Also, it will print 11, 9, and 7 reps when you need 12, 10, and 8 reps to increase weight because it mainly motivates you to get one more beyond your possibility. It's more of a mental trick.